Add files via upload
In this project, we investigate machine learning method to recognize the people using statistical data 
from laser scanner. 3 classifiers including k nearest neighbors, support vector machine and Naïve Bayes 
classifiers are used and compared with each other in the view of performance. The performance is 
measured in the aspects of TP & FP rate, precision, recall and F-measure. The experimental results 
show that the knn classifier has the best performance for the task.